requirements.txt
ffmpeg-python
whisper
torch
googletrans==4.0.0-rc1
TTS
moviepy
opencv-python
numpy
deepface
